Move SOAP encryption methods to it's own util class Make configurable to omit cert from message Make configurable to remove the soap security header

+++ core/src/test/java/org/frankframework/monitoring/MonitorTest.java
@@ -84,7 +84,7 @@ public class MonitorTest {
 		assertTrue(session.containsKey(PipeLineSession.ORIGINAL_MESSAGE_KEY));
 		Message originalMessage = (Message) session.get(PipeLineSession.ORIGINAL_MESSAGE_KEY);
 		assertEquals(message.asString(), sender.getSessionOriginalMessageValue());
-		assertEquals(123, originalMessage.getContext().get("special-key"));
+		assertEquals(123, (int) originalMessage.getContext().get("special-key"));
 	}
 
 	private String ignoreHostname(String result) {
